+++ src/fides/api/models/privacy_experience.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 from enum import Enum
 from typing import Any, Dict, List, Optional, Set, Type
-from uuid import uuid4
 
 from sqlalchemy import Boolean, Column
 from sqlalchemy import Enum as EnumColumn
@@ -22,43 +21,6 @@ from fides.api.models.location_regulation_selections import PrivacyNoticeRegion
 from fides.api.models.privacy_notice import PrivacyNotice
 from fides.api.models.property import Property
 from fides.api.schemas.language import SupportedLanguage
-
-
-class ExperienceNotices(Base):
-    """
-    A many-to-many table that links Privacy Notices to shared Privacy Experience Configs.
-    """
-
-    def generate_uuid(self) -> str:
-        """
-        Generates a uuid with a prefix based on the tablename to be used as the
-        record's ID value
-        """
-        try:
-            prefix = f"{self.current_column.table.name[:3]}_"  # type: ignore
-        except AttributeError:
-            prefix = ""
-        uuid = str(uuid4())
-        return f"{prefix}{uuid}"
-
-    # Overrides Base.id so this is not a primary key.
-    # Instead, we have a composite PK of notice_id and experience_config_id
-    id = Column(String(255), default=generate_uuid)
-
-    notice_id = Column(
-        String,
-        ForeignKey("privacynotice.id", ondelete="CASCADE"),
-        index=True,
-        nullable=False,
-        primary_key=True,
-    )
-    experience_config_id = Column(
-        String,
-        ForeignKey("privacyexperienceconfig.id", ondelete="CASCADE"),
-        index=True,
-        nullable=False,
-        primary_key=True,
-    )
 
 
 class ComponentType(Enum):
@@ -209,7 +171,7 @@ class PrivacyExperienceConfig(PrivacyExperienceConfigBase, Base):
     privacy_notices: RelationshipProperty[List[PrivacyNotice]] = relationship(
         "PrivacyNotice",
         secondary="experiencenotices",
-        back_populates="experience_configs",
+        backref="experience_configs",
         lazy="selectin",
     )
